Split project in separate files and remove unused files.
